Add the NPC class The code is all refactored from the Entity class without edits to functionality.
